bolt-simple-taproot: define initial version of taproot channels
In this extension BOLT, we specify the initial flavor of taproot
channels to be deployed. This channel type uses musig2 aggregated keys
and signatures for the funding output, making it a normal single
signature key path spend. All outputs are then updated to use P2T2
(segwit v1) outputs. The coop close process has been simplified to
always terminate, and the co-op close transaction now also flags RBF to
make way for future schemes that enable the process to be restarted
which enables co-op close fee bumping.

A top-level key spend output is used to the revocation of HTLC outputs.
The revocation for the local output uses a script path to ensure that
information needed to sweep the anchors by 3rd parties is always
revealed on chain.