temp - dyn: allow changing `channel_flags`
This allows a private/public channel to be changed to public/private.
Changing from being private to public means the participants need to
create new gossip msg set and announce them.
Changing from being public to private has limited effect as the channel
point has already been announced to the public, but a new
`channel_update` should be created and sent to the public network to
mark this channel as disabled. This would further allow an existing
public channel to be upgraded to a simple taproot channel.